requirements.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
codecov==2.1.12
coverage==7.2.2
Django==4.1.7
exceptiongroup==1.1.1
idna==3.4
iniconfig==2.0.0
packaging==23.0
Pillow==9.5.0
pluggy==1.0.0
pytest==7.2.2
pytest-cov==4.0.0
pytest-django==4.5.2
python-decouple==3.8
requests==2.28.2
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.15
